Remove ClientConectionContext.WriteMessageAsync
